Enhancements in display of attributes and datasets.
Converted attribute widget into a composite widget and nested TableView
inside. Added node path and filepath as selectable labels. This adds the
ability to copy the paths (a feature needed but missing in older
dataviz).

Added the node path and filename to the tooltip text for both dataset
widget and attribute widget.