Updated notes. Added initial DirReader class.
DirReader is intended to be widget for defining rules for loading data
organized in a directory structure. It should allow user to select a
base directory and then the rest of the path components can be defined
by rules (like "date(YYYY-mm-dd)", or "*" etc). Similarly, file name
patterns can be specified. The configurations can be named and saved in
QSettings. The ultimate goal is to automatically read all the files into
datastructures (neo?) and allow plotting and analysis inside dataviz.

At some point this can provide a way to convert such data into an HDF5
format.